// _sdk-src/src/interfaces/IStpConnector.ts
/**
 * STP connection interface
 * @interface
 */
export interface IStpConnector {
    /**
     * Name of the service connected to STP
     */
    baseName?: string;
  
  /**
   * Unique service instance name, different if there are concurrent instances running
   */
  name: string | undefined;

  /**
   * True if the connection is open and capable of sending and receiving messages
   */
  isConnected: boolean;

  /**
   * Connect and register the service, informing of the subscriptions it handles / consumes
   * @param serviceName - Name of the service that is connecting
   * @param solvables - Array of messages that this service handles
   * @param timeout - Optional number of seconds to wait for a connection before failing
   * @param machineId - Optional machine Id to use. If not provided, it is set to some unique Id.
   * @param sessionId - Optional session Id to use. If not provided:
   *  1. the suffix to the WebSocket connection string is used 
   *  2. if no WebSocket suffix was provided, the machineId is used.
   * @returns The actual sessionId used, which may be relevant if it was generated by the connector
   */
  connect(
    serviceName: string,
    solvables: string[],
    timeout?: number,
    machineId?: string | null,
    sessionId?: string | null
  ): Promise<string | undefined>;

  /**
   * Disconnect from STP
   */
  disconnect(timeout?: number): Promise<void>;

  /**
   * Send a message/command to STP
   * @param message - STP API message to send
   */
  inform(message: string, timeout?: number): Promise<void>;
  /**
   * Make a STP request - equivalent to an RPC call
   * @param message - STP API message to send
   * @returns STP API responses
   */
  request(message: string, timeout?: number): Promise<any>;

  /**
   * Event handler invoked by STP when a message matching one of the Solvables is posted by some service
   * @param message - STP API message to handle
   */
  onInform: ((message: string) => void) | undefined;
  /**
   * Event handler invoked by STP when a service makes a request matching one of the Solvables
   * @param message - STP API message to handle
   * @returns STP API response
   */
  onRequest: ((message: string) => string[]) | undefined;
  /**
   * Event handler invoked when a connection error occurs
   * @param error - Error description
   */
  onError: ((error: string) => void) | undefined;
}

/**
 * STP connection construction interface
 * @interface
 */
export interface IStpConnectorConstructor {
  /**
   * Create an STP  connection object
   * @param connstring - Connection string - may vary depending on the requirements of the connector
   */
  new (connstring: string): IStpConnector;
}